run generator on prep
