+++ www/apps/book/package.json
@@ -38,6 +38,7 @@
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react-hooks": "^5.0.0",
     "postcss": "^8",
+    "tags": "*",
     "tailwind": "*",
     "tailwindcss": "^3.3.0",
     "tsconfig": "*",
